requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
alembic>=1.14.0
scikit-learn>=1.6.0
joblib>=1.4.2
prometheus-fastapi-instrumentator>=7.0.0
